add tests

+++ product.go
@@ -7,9 +7,9 @@ import (
 
 type (
 	Product struct {
-		ID    int64  `json:"ID"`
-		Name  string `json:"name"`
-		Price int32  `json:"price"`
+		ID    int64   `json:"ID"`
+		Name  string  `json:"name"`
+		Price float32 `json:"price"`
 	}
 
 	ProductRepo interface {
